\name{jekoChr11Array250Knsp}
\docType{data}
\alias{jekoChr11Array250Knsp}
\title{Affymetrix GeneChip Mapping 250K NSP Array data of JEKO-1 cell line (chr. 11)}
\description{
  Affymetrix GeneChip Mapping 250K NSP Array data of JEKO-1 cell line.}
\usage{data(jekoChr11Array250Knsp)}
\format{A data frame containing four variables: first is SNP name ('SNPname'),
  second is probe chromosome ('Chromosome'), third is probe position ('PhysicalPosition') and fourth is probe raw log2ratio ('log2ratio').}
\source{
Poretti, G. Rancoita, P.M.V. Kwee, I. Bertoni, F., unpublished}
\keyword{datasets}
