\name{rec10k}
\docType{data}
\alias{rec10k}
\title{Affymetrix GeneChip Mapping 10K Array data of REC-1 cell line}
\description{
  Affymetrix GeneChip Mapping 10K Array data of REC-1 cell line taken
  from the reference below.
}
\usage{data(rec10k)}
\format{A data frame containing five variables: first is SNP name ('SNPname'),
  second is probe chromosome ('Chromosome'), third is probe position ('PhysicalPosition'), fourth is probe raw log2ratio ('log2ratio') and fifth
  are is probe genotype ('call').}
\source{
Rinaldi et al. (2006), Genomic and expression profiling identifies the B-cell associated tyrosine kinase Syk as a possible therapeutic
target in mantle cell lymphoma, \emph{British Journal of Haematology}, 132, 303-316}
\keyword{datasets}
